* PIECEVENTEECHEANCE.cs: fix reglement echeance
* RestServices.cs:
* REGLEMENTECHEANCE.cs:

+++ PFE/PFE/Models/PIECEVENTEECHEANCE.cs
@@ -3,14 +3,14 @@ namespace PFE.Models
 {
     public class PIECEVENTEECHEANCE
     {
-        public float? PCVID { get; set; }
-        public float? PEVID { get; set; }
+        public int? PCVID { get; set; }
+        public int? PEVID { get; set; }
         public DateTime? PEVDATE { get; set; }
-        public float? PEVMONTANT { get; set; }
-        public float? PEVTAUX { get; set; }
-        public float? RGTID { get; set; }
+        public int? PEVMONTANT { get; set; }
+        public int? PEVTAUX { get; set; }
+        public int? RGTID { get; set; }
         public string PEVISREGLE { get; set; }
-        public float? PEVMNTREGLE { get; set; }
+        public int? PEVMNTREGLE { get; set; }
         public string PITCODE { get; set; }
     }
 }
